test(PromiseBackend): convert PromiseBackend to factory class instead of singleton
Previously, some properties were being set on the PromiseBackend statically at initialization time,
which left the backend polluted between tests. This included the global object, the placeholder for 
__OriginalPromise__, and the queue of tasks. Now, these items are created in the constructor of the
PromiseBackend. Although the queue is created inside the constructor, it is attached statically to the PromiseBackend class, and is shared for all Promise/PromiseBackend instances. The only other static member of the PromiseBackend class is the executeAsap function, which is used inside of PromiseMock.